+++ doc/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Made help commands respond in private message to reduce chat spam
 - Added target lower modifier (tl) 
+- Added reroll greater (rg) and reroll greater indefinite (irg) modifiers
 - Added support for Cyberpunk RED
 - Added support for Witcher d10
 - Added support for additional wrath dice for wrath and glory